Rework the stack, options, and state.
Correctly implement the handling of ResetState when invoked via the keyboard.
Add flags to configure the header and text font sizes.
Use a constant value for the stack_depth rather than a dynamically
reallocating the stack array.
Rename the Unknown instruction to Unsupported, some of te instructions
passed in via the rom files are valid, just not supported in this
implementation.